requirements.txt
Cerberus
Eve
Events
Flask
Flask-Bootstrap
Jinja2
MarkupSafe
Pygments
Sphinx
Werkzeug
argparse
docutils
httpretty
itsdangerous
mongokit
nose
py-bcrypt
pymongo
python-dateutil
requests
six
tornado
urllib3
wsgiref
